requirements.txt
# Core framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML & Data Processing (esencial)
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0

# HTTP requests
httpx>=0.25.0
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.8.0
